+++ src/main/java/com/api/nuvu/NuvuAPI/dao/api/IPersonDAOAPI.java
@@ -3,5 +3,5 @@ package com.api.nuvu.NuvuAPI.dao.api;
 import com.api.nuvu.NuvuAPI.model.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IPersonDAOAPI extends CrudRepository<Person, Long> {
+public interface IPersonDAOAPI extends CrudRepository<Person, Integer> {
 }
