+++ src/main/java/com/api/nuvu/NuvuAPI/service/api/IPersonServiceAPI.java
@@ -3,5 +3,5 @@ package com.api.nuvu.NuvuAPI.service.api;
 import com.api.nuvu.NuvuAPI.commons.IGenericServiceAPI;
 import com.api.nuvu.NuvuAPI.model.Person;
 
-public interface IPersonServiceAPI extends IGenericServiceAPI<Person, Long> {
+public interface IPersonServiceAPI extends IGenericServiceAPI<Person, Integer> {
 }
